UI layout changed + heatmap change
UI is now along the bottom instead of the side and changes based on if
you are connected or not.
Heatmaps now only save when you close the application. This is due to it
hanging the program for a few milliseconds meaning while you are
drawing, there are small gaps due to the draw client not reading mouse
position.